e2e: Decouple logic for running e2e tests
